requirements.txt
astroid==1.4.9
dj-database-url==0.4.1
Django==1.10.4
django-admin-sortable==2.0.21
django-braces==1.10.0
django-choices==1.4.4
django-cors-headers==1.3.1
django-extensions==1.7.5
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-polymorphic==1.0.2
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.1
inflection==0.3.1
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.3
oauthlib==1.1.2
psycopg2==2.6.2
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
six==1.10.0
wrapt==1.10.8
